refactor: user와 article 패키지 분리
