+++ src/main/java/kr/codesqaud/cafe/repository/article/ArticleRepository.java
@@ -1,4 +1,4 @@
-package kr.codesqaud.cafe.repository;
+package kr.codesqaud.cafe.repository.article;
 
 import kr.codesqaud.cafe.domain.Article;
